refactor: available_pairs is for webserver only
